Revert "navigointi kaikkialle, toimii"
