+++ choresapp/App.js
@@ -1,103 +1,60 @@
-import 'react-native-gesture-handler';
-
 import React from 'react';
 import {
-  Button, 
+  SafeAreaView,
+  ScrollView, 
+  StatusBar,
+  StyleSheet,
+  Text,
+  useColorScheme,
+  Image,
   View,
-  Text
 } from 'react-native';
 
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
-
-//import { NavigationContainer } from '@react-navigation/native';
-//import { createNativeStackNavigator } from '@react-navigation/native-stack';
-
+import Header from './components/Header';
 import Adult1 from './screens/Adult1';
-import Child1 from './screens/Child1'; 
-import FirstPage from './screens/FirstPage';
-import Adult2 from './screens/Adult2';
-import Child2 from './screens/Child2';
+import Child1 from './screens/Child1';
 
-const Stack = createStackNavigator();
-
-function App() {
+const App = ()=> {
   return (
-    <NavigationContainer>
-      <Stack.Navigator initialRouteName="FirstPage">
-      <Stack.Screen
-          name="Home"
-          component={FirstPage}
- //         options={{
- //           title: 'First Page', //Set Header Title
- //           headerStyle: {
- //             backgroundColor: '#f4511e', //Set Header color
- //           },
- //           headerTintColor: '#fff', //Set Header text color
- //           headerTitleStyle: {
- //             fontWeight: 'bold', //Set Header text style
- //           },
- //         }}
-        />
-        <Stack.Screen
-          name="Adult1"
-          component={Adult1}
-          options={{
-            title: 'Adult view', //Set Header Title
-            headerStyle: {
-              backgroundColor: 'lightpink', //Set Header color
-            },
-            headerTintColor: 'darkgreen', //Set Header text color #fff white
-            headerTitleStyle: {
-              fontWeight: 'bold', //Set Header text style
-            },
-          }}
-        />
-        <Stack.Screen
-          name="Child1"
-          component={Child1}
-          options={{
-            title: 'Child view', //Set Header Title
-            headerStyle: {
-              backgroundColor: 'lightpink', //Set Header color
-            },
-            headerTintColor: 'darkgreen', //Set Header text color
-            headerTitleStyle: {
-              fontWeight: 'bold', //Set Header text style
-            },
-          }}
-        />
-        <Stack.Screen
-          name="Adult2"
-          component={Adult2}
-          options={{
-            title: 'Adult view', //Set Header Title
-            headerStyle: {
-              backgroundColor: 'lightpink', //Set Header color
-            },
-            headerTintColor: 'darkgreen', //Set Header text color
-            headerTitleStyle: {
-              fontWeight: 'bold', //Set Header text style
-            },
-          }}
-        />
-        <Stack.Screen
-          name="Child2"
-          component={Child2}
-          options={{
-            title: 'Child view', //Set Header Title
-            headerStyle: {
-              backgroundColor: 'lightpink', //Set Header color
-            },
-            headerTintColor: 'darkgreen', //Set Header text color
-            headerTitleStyle: {
-              fontWeight: 'bold', //Set Header text style
-            },
-          }}
-        />
-      </Stack.Navigator>
-    </NavigationContainer>
-  )
+    <View style={StyleSheet.container}>
+      <Header title={'Chores App'} />
+      <Text>Aikuinen</Text>
+      <View style={styles.imageContainer}>
+      <Image
+      fadeDuration={5000}
+      source={{uri:'https://static.vecteezy.com/system/resources/previews/004/477/337/non_2x/face-young-man-in-frame-circular-avatar-character-icon-free-vector.jpg'}}
+        style={styles.image} resizeMode='cover'/>
+      </View>
+      <Text>Lapsi</Text>
+      <View style={styles.imageContainer}>
+      <Image
+      fadeDuration={5000}
+      source={{uri:'https://img.freepik.com/premium-vector/avatar-portrait-kid-caucasian-boy-round-frame-vector-illustration-cartoon-flat-style_551425-43.jpg?w=2000'}}
+        style={styles.image} resizeMode='cover'/>
+      </View>
+    </View>
+  );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: 'white',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+
+  imageContainer: {
+    height: 200,
+    width: '50%',
+    borderRadius:170,
+    overflow: 'hidden',
+    borderWidth:2,
+    borderColor:'black',
+  },
+  image: {
+    height: '100%',
+    width:'100%'
+  }
+});
 
 export default App;
