+++ choresapp/package.json
@@ -15,18 +15,11 @@
     "@react-navigation/stack": "^6.3.2",
     "react": "18.1.0",
     "react-native": "0.70.2",
-<<<<<<< HEAD
-    "react-native-gesture-handler": "^2.7.1",
-    "react-native-reanimated": "^2.11.0",
-    "react-native-safe-area-context": "^4.4.1",
-    "react-native-screens": "^3.18.1"
-=======
     "react-native-gesture-handler": "^2.7.0",
     "react-native-reanimated": "^2.10.0",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.0",
     "react-native-sqlite-storage": "^6.0.1"
->>>>>>> b80162d3a3910d658dfd788ff1c0143e6a91aac5
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
